requirements.txt
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
packaging==24.0
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.2.0
pycparser==2.21
pypdf==4.1.0
PyPDF2==3.0.1
pypdfium2==4.28.0
pytesseract==0.3.10
